Include changes to classifier and a custom container.
Some changes were made to the classifier and the CAE such that it is more fitting for the
current workload.

Created a custom container with the dependencies used in the project, that way I don't have
to instantiate a new container every time and rerun the commands to install them.